directly declare base values in activation, without an interpreter
